refactor: PathRequest 생성 및 서비스 레이어에서 dto-domain 변환하도록 수정

+++ src/main/java/wooteco/subway/dto/PathResponse.java
@@ -1,16 +1,17 @@
 package wooteco.subway.dto;
 
-import wooteco.subway.domain.Path;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import wooteco.subway.domain.Path;
 
 public class PathResponse {
 
     private final List<StationResponse> stations;
-    private final double distance;
+    private final int distance;
     private final int fare;
 
-    public PathResponse(final List<StationResponse> stations, final double distance, final int fare) {
+    public PathResponse(final List<StationResponse> stations, final int distance, final int fare) {
         this.stations = stations;
         this.distance = distance;
         this.fare = fare;
@@ -20,7 +21,7 @@ public class PathResponse {
         final List<StationResponse> stationResponses = path.getStations().stream()
                 .map(StationResponse::from)
                 .collect(Collectors.toList());
-        final double distance = path.getDistance();
+        final int distance = path.getDistance();
         final int price = path.getFare().getPrice();
 
         return new PathResponse(stationResponses, distance, price);
@@ -30,7 +31,7 @@ public class PathResponse {
         return stations;
     }
 
-    public double getDistance() {
+    public int getDistance() {
         return distance;
     }
